Refactor DCT into policy/eda versions
